Add files via upload
Our project looks into predicting customer attrition using various classification methods on a bank’s customer data. We test various logistic regression, k-nearest neighbors, and random forest models to see how well each model works and to find out which model works best. Our analysis found that the random forest was the most powerful predictor of customer attrition, and the most important features were various indicators of customer engagement with the bank’s products.
Introduction and Problem Statement
Customer attrition is the loss of customers by a business. Most customers of a given business will not remain active customers indefinitely, and understanding the reason behind attrition may be beneficial for a business in the long term to increase retention. Customers are valuable to a business, as not only are they the source of a business’ revenue, they are expensive and difficult to get. According to DCR strategies, monitoring attrition rates can help businesses save money on marketing, increase customer lifetime value, and get free word-of-mouth advertising.
In this project, our team is using the Credit Card customers dataset from Kaggle. Our goal is to help this bank figure out which customers are most probable to leave. This will help them direct their attention to those customers and provide better services. 
Our success metric is our ability to find customers who are likely to leave. We will compare our different models by their recall (proportion of correctly identified attriting customers out of all actual attriting customers) and the area under the curve (AUC) of their receiver operating characteristic (ROC) curves, which indicates the ability to balance the true positive rate and false positive rate.
Final Model Selection
Based on each model’s recall and AUC, we see that the random forest model performed best, even when tuned on just a handful of key features. This lines up with our expectations, as we know random forests to be one the most powerful classification algorithms. As this was our best model we use it for our variable interpretations.
Interpretations
To better understand the relationship between the key variables in our final model and churn rates, we plotted SHAP independent graphs with the top four variables, which indicate a variable’s effect on the model’s decision. From the graph below and on the left, we can see that customers who make more transactions are more likely to keep their credit card. For the transaction amount, we can spot that customers are more likely to keep their card when their transaction amount in the past 12 months is around either $1,250 or $4,000. 
Conclusion
Our final model performed quite well using only a few variables, indicating that these variables were key predictors of customer churn. Four of these variables (transaction amount and count, revolving balance, utilization rate) indicated how much the customers were using their card(s), two more (count/amount change from Q1 to Q4) indicated how much their spending increased/decrease over time, and the seventh (relationship count) indicated how many of the bank’s products the customer used. All of these variables represent a customer’s engagement with the bank, and in most cases, more engagement meant more retention. 
However, this brings up a key limitation with our project. The data does not specify when the customers attrited, and how the data was collected for attrited customers. As such, we do not know if the metrics such as total spending in the past 12 months was lower for attrited customers because of their behavior or because they left during those 12 months, and therefore were not present for later months. As this is just a project on found data, we cannot interact with the bank’s database directly. In a real business scenario, this would be a simple issue of collecting accurate data. For this project, we can simply treat our previous analysis as the framework by which a bank would try to predict attrition.
A second, smaller limitation of the data is that the nature of the key features requires the bank to measure a customer’s behavior over time to predict whether or not they will churn. While some variables can be approximated - namely, spending in the past 12 months can be extrapolated from a customer’s spending in the past three months or so - others cannot - such as measuring the change in quarterly spending, as this looks for deviations from normal seasonal trends. While this project can be used to help the bank identify customers who may be on their way out, it cannot predict whether new customers will churn in the future. Such purposes require more research and perhaps more information about the customer’s demographic or financial data. Also, if a bank wishes to apply this model to its customers, it must be careful to make sure their metrics for that customer accurately represent them the way the model expects.
In terms of positives, there are strong future possibilities for our analysis. Banks looking to retain their customers for longer can use our models to find customers who are losing interest, and apply new marketing strategies targeting those customers, then also use our models to measure the impact of those marketing strategies. In an A/B test, a firm could compare the attrition rate in the treatment group not just to that of the control group, but to the treatment group’s predicted attrition rate generated by our model.
Further iterations of this analysis could potentially find customers who are not only likely to churn, but likely to leave regardless of additional incentives to stay. If a bank were reliably able to identify such customers, they could save resources by letting these customers go, as well remind them to pay off their credit card debt to improve their credit scores, so the bank ensures they get as much customer value as possible.